requirements.txt
google-adk==1.18.0
firebase_admin
google-cloud-firestore
fastapi
uvicorn
